Various changes. Added Spatdif Handler.

+++ include/decoder.h
@@ -29,7 +29,7 @@ public:
   virtual void Decode() = 0;
   
   MonoStream* stream(Int chan_id) {
-    if (chan_id < 0 || chan_id > output_streams_.size()) { throw_line(""); }
+    if (chan_id < 0 || chan_id > (Int) output_streams_.size()) { throw_line(""); }
     return &(output_streams_.at(chan_id));
   }
